requirements.txt
pandas==1.0.4
plotly==4.8.2
numpy==1.18.4
streamlit==0.60.0
matplotlib==3.2.1
Pillow==7.2.0
scikit_learn==0.23.1